+++ lib/livesecret.ex
@@ -7,24 +7,19 @@ defmodule LiveSecret do
   if it comes from the database, an external API or others.
   """
 
-  alias LiveSecret.{Do, PubSubDo, Secret}
+  alias LiveSecret.{Do, PubSubDo}
 
   defdelegate count_secrets(tenant), to: Do
   defdelegate get_secret!(tenant, id), to: Do
   defdelegate get_secret(tenant, id), to: Do
   defdelegate insert!(tenant, presecret_attrs), to: Do
   defdelegate validate_presecret(tenant, presecret_attrs), to: Do
-
-  def burn!(secret, event_extra \\ []) do
-    secret = %Secret{id: id, burned_at: burned_at} = Do.burn!(secret)
-    PubSubDo.notify_burned!(id, burned_at, event_extra)
-    secret
-  end
+  defdelegate watch_secret(tenant, label, id), to: Do
+  defdelegate burn!(secret), to: Do
 
   defdelegate go_live!(tenant, id), to: Do
   defdelegate go_async!(tenant, id), to: Do
 
   defdelegate subscribe!(id), to: PubSubDo
   defdelegate notify_unlocked!(id, user_id), to: PubSubDo
-  defdelegate notify_expired(id), to: PubSubDo
 end
