+++ lib/livesecret/pubsub_do.ex
@@ -16,25 +16,4 @@ defmodule LiveSecret.PubSubDo do
   def notify_unlocked!(id, user_id) do
     :ok = Phoenix.PubSub.broadcast(LiveSecret.PubSub, Secret.topic(id), {:unlocked, user_id})
   end
-
-  @doc """
-  Notifies PubSub topic for the secret that the given user `burned_by` burned the secret at
-  timestamp `burned_at`. All listeners should update their state for this secret.
-  """
-  def notify_burned!(id, burned_at, extra) do
-    :ok =
-      Phoenix.PubSub.broadcast(
-        LiveSecret.PubSub,
-        Secret.topic(id),
-        {:burned, burned_at, extra}
-      )
-  end
-
-  @doc """
-  Notifies PubSubtopic for the secret that it has been expired. All listeners should update
-  their state for this secret.
-  """
-  def notify_expired(id) do
-    Phoenix.PubSub.broadcast(LiveSecret.PubSub, Secret.topic(id), :expired)
-  end
 end
